+++ packages/web3-providers/tests/src/providers/HttpProviderTest.js
@@ -289,15 +289,44 @@ describe('HttpProviderTest', () => {
     it('calls sendPayload and returns with a rejected promise because of the exceeded timeout', async () => {
         new XHR();
         const xhrMock = XHR.mock.instances[0];
+        xhrMock.addEventListener = jest.fn((event, callback) => {
+            if (event === 'timeout') {
+                callback();
+            }
+        });
 
         providersModuleFactoryMock.createXMLHttpRequest.mockReturnValueOnce(xhrMock);
 
-        setTimeout(() => {
-            xhrMock.ontimeout();
-        }, 1);
-
         await expect(httpProvider.sendPayload({id: '0x0'})).rejects.toThrow(
             'Connection error: Timeout exceeded after 1ms'
+        );
+
+        expect(httpProvider.connected).toEqual(false);
+
+        expect(providersModuleFactoryMock.createXMLHttpRequest).toHaveBeenCalledWith(
+            httpProvider.host,
+            httpProvider.timeout,
+            httpProvider.headers,
+            httpProvider.agent,
+            httpProvider.withCredentials
+        );
+
+        expect(xhrMock.send).toHaveBeenCalledWith('{"id":"0x0"}');
+    });
+
+    it('calls sendPayload and returns with a rejected promise because of the http error listener', async () => {
+        new XHR();
+        const xhrMock = XHR.mock.instances[0];
+        xhrMock.addEventListener = jest.fn((event, callback) => {
+            if (event === 'error') {
+                callback();
+            }
+        });
+
+        providersModuleFactoryMock.createXMLHttpRequest.mockReturnValueOnce(xhrMock);
+
+        await expect(httpProvider.sendPayload({id: '0x0'})).rejects.toThrow(
+            `Network error ${JSON.stringify({id: '0x0'})}`
         );
 
         expect(httpProvider.connected).toEqual(false);