+++ CHANGELOG.md
@@ -90,3 +90,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [2.0.0-alpha.2]
+
+### Fixed
+
+- Error handling of the HttpProvider fixed (#2887)
+